# api/openapi.yml
openapi: 3.0.0
info:
  title: User Service API
  version: 1.0.0
paths:
...
  /tasks/{task_id}:
    get:
      summary: get task by id
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
        '401':
          description: Invalid or expired token
        '500':
          description: Internal error
    put:
      summary: Update a task
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
        - in: query
          name: title
          required: true
          schema:
            type: string
        - in: query
          name: content
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
        '401':
          description: Invalid or expired token
        '500':
          description: Internal error
    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '401':
          description: Invalid or expired token
        '500':
          description: Internal error
...
  /top-users:
    get:
      summary: Top 3 authors by likes
      responses:
        '200':
          description: Top 3 authors
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                    total_likes:
                      type: integer
        '401':
          description: Invalid or expired token
        '500':
          description: Internal error